+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 * Provide a route delegate.
 * Enable inner navigation.
 * Enable navigation stack history saved.
+
+## 0.0.2
+
+* Added ScopedPageBuilder which provide functionality to get notified when initial page has been exited.
+* Added widget tests that cover the new ScopedPageBuilder functionality.
+* Renamed the JsonPojoConverter to DestinationArgumentConverter to unify the api and make it more clear.